agrego la posibilidad de modificar el estado del pedido a finalizado

+++ Backend/ecommerce/src/main/java/com/example/ecommerce/service/ProductService.java
@@ -143,7 +143,7 @@ public class ProductService implements IProductService {
 
     public void setOfferAvailability(List<Product> products) {
         for (Product product : products) {
-            if (product.getOffer().getEndOffer().after(new Date())) {
+            if (product.getOffer() != null && product.getOffer().getEndOffer().after(new Date())) {
                 iOfferRepository.delete(product.getOffer());
             }
         }
